% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hf_utilities.R
\name{add_directorate_variable}
\alias{add_directorate_variable}
\title{add_directorate_variable}
\usage{
add_directorate_variable(table, spec_column = "main_specialty_start")
}
\arguments{
\item{table}{a dataframe with a column containing specialties}

\item{spec_column}{string containing column name containing specialties}
}
\value{
table with an additional column containing the directorate associated
with the specialty for that row.
}
\description{
add_directorate_variable
}
