% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admission_discharges.R
\name{admission_discharges}
\alias{admission_discharges}
\title{admission_discharges}
\usage{
admission_discharges(start_date, end_date, data, plot_chart)
}
\arguments{
\item{start_date}{date of earliest discharge to be included in the analysis}

\item{end_date}{date of latest admission to be included in the analysis}

\item{data}{hospital episode data}

\item{plot_chart}{if TRUE return chart, otherwise if FALSE return dataframe}
}
\value{
Chart or dataframe showing average Admissions and Discharges by the day of the week
}
\description{
admission_discharges
}
\examples{
\dontrun{
hospitalflow::admission_discharges(start_date = "2015-01-01 00:00:00", end_date = "2015-02-01 00:00:00",
                                   data = example_data, plot_chart = TRUE)
}
}
