% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admissions_discharges.R
\name{admissions_discharges}
\alias{admissions_discharges}
\title{Admission and Discharges Function
data isn't  standardised
admission_discharges}
\usage{
admissions_discharges(
  start_date,
  end_date,
  data,
  plot_chart,
  hospital_name = "Hospial name"
)
}
\arguments{
\item{start_date}{date of earliest discharge to be included in the analysis}

\item{end_date}{date of latest admission to be included in the analysis}

\item{data}{hospital episode data with at least the following fields:
Admissions - the admission date
Discharges - the discharge date
IDcol - the patient pseudo id
PatientType - the admission type, distinguishing emergency from other admissions
EpisodeNumber - the episode id
WardCode - the ward for the episode}

\item{plot_chart}{if TRUE return chart, otherwise if FALSE return dataframe}
}
\value{
Chart or dataframe showing average Admissions and Discharges by the day of the week
}
\description{
Admission and Discharges Function
data isn't  standardised
admission_discharges
}
\examples{
\dontrun{
hospitalflow::admissions_discharges(start_date = "2015-01-01 00:00:00", end_date = "2015-02-01 00:00:00",
                                   data = example_data, plot_chart = TRUE)
}
}
