% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrival_occupancy.R
\name{ae_arrival_occupancy_fct}
\alias{ae_arrival_occupancy_fct}
\title{ae_arrival_occupancy_fct}
\usage{
ae_arrival_occupancy_fct(start_date = as.Date("2012-01-01", tz =
  "Europe/London"), end_date = as.Date("2015-01-01", tz =
  "Europe/London"), data, plot_chart, hospital_name = "Hospital name")
}
\arguments{
\item{hospital_name}{}
}
\description{
ae_arrival_occupancy_fct
}
