% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_attendances_admissions_age_sex.R
\name{ae_attendances_admissions_age_sex}
\alias{ae_attendances_admissions_age_sex}
\title{ae_attendances_admissions_age_sex}
\usage{
ae_attendances_admissions_age_sex(start_date = as.Date("2012-01-01", tz =
  "Europe/London"), end_date = as.Date("2015-01-01", tz = "Europe/London"),
  data, plot_chart, hospital_name = "Chelsea & Westminster")
}
\arguments{
\item{start_date}{date of earliest discharge to be included in the analysis}

\item{end_date}{date of latest admission to be included in the analysis}

\item{data}{hospital episode data}

\item{plot_chart}{if TRUE return chart, otherwise if FALSE return dataframe}

\item{hospital_name}{this is a string specifying the hospital name for the chart title}
}
\value{
Chart or dataframe showing A&E Attendances and Admissions by Gender and Age
}
\description{
ae_attendances_admissions_age_sex
}
\examples{
\dontrun{
hospitalflow::ae_attendances_admissions_age_sex(start_date = "2012-01-01 00:00:00",
end_date = "2015-01-01 00:00:00",
data = test_data_age_sex_att_adm, plot_chart = TRUE)
}
}
