% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_attendances_admissions_age_sex.R
\name{age_sex}
\alias{age_sex}
\title{age_sex}
\usage{
age_sex(config_path = "lgt-config/ed")
}
\arguments{
\item{config_path}{path to the ed config folder}
}
\value{

}
\description{
age_sex
}
