% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrival_occupancy.R
\name{arrival_occupancy}
\alias{arrival_occupancy}
\title{arrival_occupancy}
\usage{
arrival_occupancy(start_date = as.Date("2012-01-01", tz = "Europe/London"),
  end_date = as.Date("2015-01-01", tz = "Europe/London"), data, plot_chart,
  hospital_name = "Chelsea & Westminster")
}
\arguments{
\item{hospital_name}{}
}
\description{
arrival_occupancy
}
