% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_standardise.R
\name{bind_into_facilities}
\alias{bind_into_facilities}
\title{bind_into_facilities}
\usage{
bind_into_facilities(data_config_list)
}
\arguments{
\item{data_config_list}{}
}
\value{
list of as many tibbles as there are unique values of facility in the provided config list. Each
tibble is formed by row binding all imported data for that facility.
}
\description{
bind_into_facilities
}
