% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hf_utilities.R
\name{convert_rds_cols_to_char}
\alias{convert_rds_cols_to_char}
\title{convert_rds_cols_to_char}
\usage{
convert_rds_cols_to_char(path)
}
\arguments{
\item{path}{}
}
\value{
list of NULL (as per saveRDS)

Loads in all RDS files in path, assumes all are dataframes. Converts all columns
to character, and re-saves as same filenames.

Use with care!
}
\description{
convert_rds_cols_to_char
}
