% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ed_data_example}
\alias{ed_data_example}
\title{Episode data for Lewisham and Greenwhich Hospital}
\format{A tibble containing observations of 16 variables. Each observation corresponds to an episode of
emergency department care.
\describe{
  \item{pseudo_id}{Unique ID number, character type, of the patient/ Original mapping back to hospital number remains
  behind Lewisham and Greenwhich NHS Trust firewall}
  \item{start_datetime}{The date that the patient attended the ED department for the spell at the hospital that
  this episode of care is part of. The type of the datetime as POSIXct}
  \item{end_datetime}{The time the patient was discharged from ED department}
  \item{gender}{A factor variable with 3 levels: "Female", "Male", "Other"}
  \item{age_band_start}{Factor variable with 22 categories. Categories formed are within bands
  of 0, 1-4, 5-9 ..... 90-94, 95+}
  \item{facility}{A factor variable ... }
  \item{ethnic_category}{The ethnicity of the patient recorded as per NHS data dictionary
  coding system using the initials for each ethicity.
  \url{https://www.datadictionary.nhs.uk/data_dictionary/attributes/e/end/ethnic_category_code_de.asp}}
  \item{attendance_Category}{The factor variable attendance category has three levels:
  "First Attendance", "Follow-up Planned", "Follow-up unplanned"}
  \item{triage_category}{The triage category factor variable recorded as per NHS data dictionary coding system
  \url{https://www.datadictionary.nhs.uk/data_dictionary/attributes/a/a_and_e_initial_assessment_triage_category_de.asp?shownav=1}}
  \item{diagnosis_string}{A character variable  with a coding system based on NHS data dictionary?}
  \url{https://www.datadictionary.nhs.uk/web_site_content/supporting_information/clinical_coding/accident_and_emergency_diagnosis_tables.asp?shownav=1}
  \item{hrg_string}{Need to check with Tom}
  \item{postcode_sector}{The postcode character variable of the patient's home, up to sector level}
  \item{arrival_mode}{The means of arrival of the patient at the ED department, variable recorded as factor.}
  \item{attendance_disposal}{Attendance disposal recorded as a factor variable, ask Tom for more info, forgotten what it means}
  \item{referal_source}{The source of referral of recorded as per NHS data dictionary coding system
  \url{https://www.datadictionary.nhs.uk/data_dictionary/attributes/s/smo/source_of_referral_for_a_and_e_de.asp?shownav=1}}
  \item{specialty}{see with Tom}
  \item{episode_id}{An interger variable calculated field. Check with Tom for a clear definition}



  }}
\usage{
ed_data_example
}
\description{
A dataset containing data ... episodes for Emergency Department(ED)
for Lewisham and Greenwhich hospital, for all patients between 2015 and 2018
}
\keyword{datasets}
