% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed_los_flgrps.R
\name{ed_los_flow_grps}
\alias{ed_los_flow_grps}
\title{ed_los_flow_grps}
\usage{
ed_los_flow_grps(start_date, end_date, data, plot_chart,
  hospital_name = "Name")
}
\arguments{
\item{hospital_name}{}
}
\description{
ed_los_flow_grps
}
