% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_modelling.R
\name{evaluate_predictions}
\alias{evaluate_predictions}
\title{evaluate_predictions}
\usage{
evaluate_predictions(predictions)
}
\arguments{
\item{predictions}{predictions tibble created by run_S_func_model}
}
\value{
object with evaluation results
}
\description{
evaluate_predictions
}
