% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_standardise.R
\name{filter_config_list}
\alias{filter_config_list}
\title{filter_config_list}
\usage{
filter_config_list(config_list, field, value)
}
\arguments{
\item{config_list}{a configuration list (of lists)}

\item{field}{the field of the inner lists to filter on}

\item{value}{the value of the field to filter to}
}
\value{
list of only those inner lists where field == value
}
\description{
filter_config_list
}
