% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ed_4hourperf_timeseries_flow.R
\name{format_control_chart}
\alias{format_control_chart}
\title{Format a control chart}
\usage{
format_control_chart(cht, r1_col, r2_col, main_col = "black")
}
\arguments{
\item{cht}{\code{ggplot} chart object.}

\item{r1_col}{Colour to highlight rule 1 breaks.}

\item{r2_col}{Colour to highlight rule 2 breaks.}

\item{main_col}{Colour without highlight.}
}
\value{
A chart with formatting applied.
}
\description{
\code{format_control_chart} applies formatting to the provided control chart.
}
\examples{
\dontrun{
TBD
}

}
