% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four_hrs_perf.R
\name{four_hrs_perf}
\alias{four_hrs_perf}
\title{four hour performance for ED}
\usage{
four_hrs_perf(start_date = as.Date("2012-01-01", tz = "Europe/London"),
  end_date = as.Date("2015-01-01", tz = "Europe/London"), data,
  plot_chart, hospital_name = "Hospital_Name")
}
\arguments{
\item{hospital_name}{}
}
\description{
four hour performance for ED
}
