% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four_hrs_perf_flow_groups.R
\name{four_hrs_perf_flow_groups}
\alias{four_hrs_perf_flow_groups}
\title{four hrs perf flow_groups}
\usage{
four_hrs_perf_flow_groups(
  start_dt = as.Date("2016-01-01"),
  end_dt = as.Date("2016-09-01"),
  data,
  time_unit = "day",
  plot_chart,
  hospital_name = "Hospital_Name",
  config_path = "lgt-config/ed"
)
}
\arguments{
\item{config_path}{path to the ed config folder}
}
\value{

}
\description{
four hrs perf flow_groups
}
