% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_standardise.R
\name{get_colname_mapping}
\alias{get_colname_mapping}
\title{get_colname_mapping}
\usage{
get_colname_mapping(config_path)
}
\arguments{
\item{config_path}{path to the config files for a given dataset}
}
\value{
character vector specifying column renaming, of the form
c(new_name_1 = "old_name_1", new_name_2 = "old_name_2")
}
\description{
get_colname_mapping
}
