% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boarding.R
\name{get_contigency_table}
\alias{get_contigency_table}
\title{get_contingency_table}
\usage{
get_contigency_table(spell_table)
}
\arguments{
\item{spell_table}{}
}
\value{

}
\description{
get_contingency_table
}
