% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boarding.R
\name{get_contingency_table}
\alias{get_contingency_table}
\title{Contingency table between ward category and main specialty/disease categories.}
\usage{
get_contingency_table(ipData, vars, scaleBy = "none", returnPlot = FALSE)
}
\arguments{
\item{ipData}{\itemize{Hospital inpatient episode data with at least the following fields:
  \item \code{ward_category} - category of the ward;
  \item \code{main_specialty} - main specialty of the clinician responsible for the episode;
  \item \code{ccsr_category_description} - disease group that ICD-10 code of the patient belongs to.
}}

\item{vars}{Variables of interest - either \code{"ws"} (ward type - main specialty) or
\code{"wd"} (ward type - CCSR disease category).}

\item{scaleBy}{\itemize{Whether and how entries of contingency table should be scaled.
  \item \code{"none"} - no scaling, presents absolute episode counts;
  \item \code{"rowsum"} - normalizes by the total number of episodes in a given ward. Useful when investigating the distribution of specialties/diseases among wards.
  \item \code{"colsum"} - normalizes by the total number of episodes in a given specialty/disease group. Useful when investigating the distribution of wards among specialties/diseases.
  \item \code{"total"} - normalizes by the total number of episodes.
}}

\item{returnPlot}{A boolean value indicating whether a plot (heatmap) should be return. If \code{FALSE}, a table is returned.}
}
\value{
A contingency table (default) or a heatmap showing the relationship between ward type and
main specialty or CCSR categories.
}
\description{
\code{get_ward_specialty_contingency_table} returns a table showing the relationship
between ward types and main specialties or disease groups. Used in patient boarding analysis to infer
what type of patients a given ward is designed to accommodate. The function is currently using
a 1-to-1 mapping between ICD-10 codes and CCSR categories. There is a possibility to
include a 1-to-many mapping in the analysis in the future if needed.
}
\examples{
\dontrun{
get_contingency_table(
  ip_data = inpatient_data,
  vars = "ws",
  scaleBy = "none",
  returnPlot = FALSE
)
}

}
