% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_standardise.R
\name{get_factor_recode}
\alias{get_factor_recode}
\title{get_factor_recode}
\usage{
get_factor_recode(config_path)
}
\arguments{
\item{config_path}{path to the directory containing the config files for this dataset}
}
\value{
copy of imported_data with factor variables recoded to standard levels as
per the config files
}
\description{
get_factor_recode
}
