% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_matrix.R
\name{get_flow_matrix}
\alias{get_flow_matrix}
\title{get_flow_matrix}
\usage{
get_flow_matrix(ed_data, inpatient_data, spell_table)
}
\arguments{
\item{spell_table}{}
}
\value{

}
\description{
get_flow_matrix
}
