% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_standardise.R
\name{get_import_col_types}
\alias{get_import_col_types}
\title{get_import_col_types}
\usage{
get_import_col_types(config_path)
}
\arguments{
\item{config_path}{path to the config files for a given dataset}
}
\value{
a column specification to be used for parsing the dataset, as a named vector of
expressions. The expressions are readr collectors, and the names specify the columns they
will parse.
}
\description{
get_import_col_types
}
