% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_modelling.R
\name{get_inpatient_snapshot}
\alias{get_inpatient_snapshot}
\title{get_inpatient_snapshot}
\usage{
get_inpatient_snapshot(df, t)
}
\arguments{
\item{df}{spell data including start_datetime and end_datetime}

\item{t}{time to take snapshot}
}
\value{
df restricted to all spells in hospital at time t
}
\description{
get_inpatient_snapshot
}
