% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_modelling.R
\name{get_poisson_binomial_characteristic_values}
\alias{get_poisson_binomial_characteristic_values}
\title{get_poisson_binomial_characteristic_values}
\usage{
get_poisson_binomial_characteristic_values(p)
}
\arguments{
\item{p}{vector of N = n + 1 probabilities, where n is number of bernoulli trials
summed to form a poisson binomial distribution}
}
\value{
values of the characteristic function of the pbd
evaluated at arguments of Nth roots of unity
}
\description{
get_poisson_binomial_characteristic_values
}
