% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_modelling.R
\name{get_residual_occupancy}
\alias{get_residual_occupancy}
\title{get_residual_occupancy}
\usage{
get_residual_occupancy(df, t, delta_t)
}
\arguments{
\item{delta_t}{}
}
\value{

}
\description{
get_residual_occupancy
}
