% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data-simulation.R
\name{get_simulated_data_as_list}
\alias{get_simulated_data_as_list}
\title{get_simulated_data_as_list}
\usage{
get_simulated_data_as_list(
  npat = 10000,
  start = as.POSIXct("2019-01-01 00:00:00"),
  end = as.POSIXct("2020-01-01 00:00:00")
)
}
\arguments{
\item{npat}{number of patients to be simulated for ED and Inpatient datasets each.
Some patients from the ED dataset will also be in the Inpatient data if they were admitted.}

\item{start}{the start datetime of the simulation period.}

\item{end}{the end datetime of the simulation period.}
}
\value{
list containing a dataframe of simulated ED data and the corresponsing Inpatient dataframe.
}
\description{
get_simulated_data_as_list
}
