% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ward_mapping.r
\name{get_ward_mapping}
\alias{get_ward_mapping}
\title{get_ward_mapping}
\usage{
get_ward_mapping(ward_name, config_path = "lgt-config/inpatient/")
}
\arguments{
\item{ward_name}{Exact ward name to use to look up corresponding specialism}

\item{config_path}{Path to config file with ward mapping}
}
\value{
The broad specialism corresponding to the input ward
}
\description{
get_ward_mapping
}
