% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boarding.R
\name{get_ward_specialty_contingency_table}
\alias{get_ward_specialty_contingency_table}
\title{Contingency table between ward category and main specialty}
\usage{
get_ward_specialty_contingency_table(ipData, scaleBy, returnPlot = FALSE)
}
\arguments{
\item{ip_data}{\itemize{Hospital inpatient episode data with at least the following fields:
  \item \code{ward_category} - category of the ward;
  \item \code{main_specialty} - main specialty of the clinician responsible for the episode;
}}
}
\value{
A contingency table showing the relationship between ward type and
main specialty.
}
\description{
\code{get_ward_specialty_contingency_table} returns a table showing the relationship
between ward types and main specialties. Used in patient boarding analysis to infer
what type of patients a given ward is designed to accommodate.
}
\examples{
\dontrun{
get_ward_specialty_contingency_table(
  ip_data = inpatient_data
)
}

}
