% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_standardise.R
\name{import_and_standardise}
\alias{import_and_standardise}
\title{import_and_standardise}
\usage{
import_and_standardise(data_import_list)
}
\arguments{
\item{data_paths}{list of paths to csv files for data import}

\item{config_paths}{list of paths to directories containing csv config files for respective element of
data_paths}
}
\value{
list of tibbles, each containing standardised import of one data file from data_paths
}
\description{
import_and_standardise
}
