% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_standardise.R
\name{import_and_standardise_inpatient}
\alias{import_and_standardise_inpatient}
\title{import_and_standardise_inpatient}
\usage{
import_and_standardise_inpatient(data_import_list)
}
\arguments{
\item{data_import_list}{list of named lists, each named list has two elements called
data_path and config_path, whose values are character strings specifying the paths
to i) a csv file to import data from and ii) a folder containing hospitalflow config
files.}
}
\value{
list of tibbles, each containing standardised import of one data file from data_paths
}
\description{
import_and_standardise_inpatient
}
