% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_standardise.R
\name{import_standardise_bind}
\alias{import_standardise_bind}
\title{import_standardise_bind}
\usage{
import_standardise_bind(data_import_list)
}
\arguments{
\item{data_import_list}{list of named lists, each named list has four elements called
data_path, config_path, site, and facility. The first two of these are character strings specifying the paths
to i) a csv file to import data from and ii) a folder containing hospitalflow config
files. The latter two specify the site and facility (e.g. ED, IP, UCC).}
}
\value{
list of as many tibbles as there are unique values of facility in the provided config list. Each
tibble is formed by row binding all imported data for that facility. N.B. when sites with different time zones
are bound together, all datetimes are converted to have the same time zone by R default.
}
\description{
import_standardise_bind
}
