% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_dataset.R
\name{lgt_col_mapping}
\alias{lgt_col_mapping}
\title{lgt_col_mapping}
\usage{
lgt_col_mapping()
}
\value{
the mapping table as a tibble
}
\description{
lgt_col_mapping
}
\examples{
lgt_col_mapping()

}
