% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/los_att_adm_ae.R
\name{los_att_adm_ae}
\alias{los_att_adm_ae}
\title{lenght of stay for attendances and admissions from ED}
\usage{
los_att_adm_ae(
  start_date = as.Date("2012-01-01"),
  end_date = as.Date("2013-01-01"),
  data,
  plot_chart,
  hospital_name = "Chelsea & Westminster",
  config_path = "lgt-config/ed"
)
}
\arguments{
\item{config_path}{path to the ed config folder}
}
\value{

}
\description{
lenght of stay for attendances and admissions from ED
}
