% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/los_distrib_maternity_other.R
\name{los_distrib_maternity_other}
\alias{los_distrib_maternity_other}
\title{length of stay for maternigy and other babies}
\usage{
los_distrib_maternity_other(start_date = as.Date("2012-01-01", tz =
  "Europe/London"), end_date = as.Date("2015-01-01", tz = "Europe/London"),
  data, plot_chart, hospital_name = "Chelsea & Westminster")
}
\arguments{
\item{start_date}{date of earliest discharge to be included in the analysis}

\item{end_date}{date of latest admission to be included in the analysis}

\item{data}{hospital episode data}

\item{plot_chart}{if TRUE return chart, otherwise if FALSE return dataframe}

\item{hospital}{name string for the hospital under analysis}
}
\value{
Chart or dataframe showing counts for Maternity and Other Babies for the 8 hrs bins to 7d, 7d bins to 28 days, and more than 28 days
}
\description{
length of stay for maternigy and other babies
}
\examples{
\dontrun{
hospitalflow::admission_discharges(start_date = "2015-01-01 00:00:00", end_date = "2015-02-01 00:00:00",
                                   data = example_data, plot_chart = TRUE)
}
}
