% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/los_distrib_method_admission.R
\name{los_distrib_method_admission}
\alias{los_distrib_method_admission}
\title{los_distrib_method_admission}
\usage{
los_distrib_method_admission(
  start_date,
  end_date,
  data,
  plot_chart,
  hospital_name = "Hospital Name"
)
}
\arguments{
\item{hospital_name}{}
}
\value{

}
\description{
los_distrib_method_admission
}
