% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_dataset.R
\name{make_adm_disch_cols}
\alias{make_adm_disch_cols}
\title{make_adm_disch_cols}
\usage{
make_adm_disch_cols(data)
}
\arguments{
\item{data}{episode data, including columns called
'Spell Number', 'START_DATETIME', 'END_DATETIME' and 'Admission.Type'}
}
\value{
data, with columns for admission and discharge datetime, and spell-level admission type (PatientType), added
}
\description{
make_adm_disch_cols
}
\examples{
\dontrun{
new_dataset <- make_adm_disch_cols(dataset)
}
}
