% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_dataset.R
\name{make_adm_type_col}
\alias{make_adm_type_col}
\title{make_adm_type_col}
\usage{
make_adm_type_col(data, adm_method_mapping)
}
\arguments{
\item{data}{episode data including ADMISSION_FROM_CODE column}

\item{adm_method_mapping}{mapping from Admission Method to type of admission. e.g. 11 = Elective.
Must have columns Admission.Method and Admission.Type, Admission.Method values should be distinct i.e.
each value of Admission.Method should have one and only one corresponding Admission.Type.}
}
\value{
data, with additional column Admission.Type mapped from ADMISSION_FROM_CODE using adm_method_mapping
}
\description{
make_adm_type_col
}
\examples{
\dontrun{
new_dataset <- make_adm_type_col(dataset, adm_method_mapping = mapping)
}
}
