% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_standardise.R
\name{make_episode_ids}
\alias{make_episode_ids}
\title{make_episode_ids}
\usage{
make_episode_ids(episode_data)
}
\arguments{
\item{episode_data}{}
}
\value{
episode_data with a new index column
}
\description{
make_episode_ids
}
