% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell_linkage.R
\name{make_moves_table}
\alias{make_moves_table}
\title{make_moves_table}
\usage{
make_moves_table(
  ed_data = test_ed_data_sample,
  inpatient_data = test_ip_data_sample,
  all_episodes,
  ward_mapping_config_path
)
}
\arguments{
\item{ed_data}{standard ED data}

\item{inpatient_data}{standard inpatient data}

\item{all_episodes}{all episodes tables (this is an output from make_spell_table)}

\item{ward_mapping_config_path}{path to the ward mapping config file}
}
\value{
moves table
}
\description{
make_moves_table
}
