% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spell_table_new.R
\name{make_spell_number_dtbl}
\alias{make_spell_number_dtbl}
\title{make_spell_number_dtbl}
\usage{
make_spell_number_dtbl(
  ed_data,
  inpatient_data,
  same_type_episode_lag = 1,
  different_type_episode_lag = 6
)
}
\arguments{
\item{ed_data}{}

\item{different_type_episode_lag}{}
}
\description{
make_spell_number_dtbl
}
