% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell_linkage.R
\name{make_spell_table}
\alias{make_spell_table}
\title{make_spell_table}
\usage{
make_spell_table(ed_data, inpatient_data)
}
\arguments{
\item{ed_data}{standard ED data}

\item{inpatient_data}{standard inpatient data}
}
\value{
spell table
}
\description{
make_spell_table
}
