% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_modelling.R
\name{make_survival_function}
\alias{make_survival_function}
\title{make_survival_function}
\usage{
make_survival_function(KM)
}
\arguments{
\item{KM}{Kaplan-Meier output of survfit}
}
\value{
the Kaplan-Meier estimate as a stepfun
}
\description{
make_survival_function
}
