% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_deaths.R
\name{measure_death}
\alias{measure_death}
\title{measure_death}
\usage{
measure_death(start_date = as.POSIXct("2016-01-01 00:00:00", tz =
  "Europe/London"), end_date = as.POSIXct("2016-03-31 00:00:00", tz =
  "Europe/London"), data, plot_chart, hospital_name)
}
\arguments{
\item{hospital_name}{}
}
\description{
measure_death
}
