% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_timeser.R
\name{mortality_timeser}
\alias{mortality_timeser}
\title{mortality_timeser}
\usage{
mortality_timeser(
  start_date = as.POSIXct("2016-01-01 00:00:00"),
  end_date = as.POSIXct("2016-03-31 00:00:00"),
  data,
  plot_chart,
  hospital_name,
  config_path = "lgt-config/ed"
)
}
\arguments{
\item{config_path}{path to the ed config folder}
}
\value{

}
\description{
mortality_timeser
}
