% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves_table.R
\docType{data}
\name{moves_table}
\alias{moves_table}
\title{Simulated moves table corresponding to ed_data and inpatient_data.}
\format{

}
\usage{
moves_table
}
\description{
A dataset containing the output of make_moves_table() when applied to ed_data and inpatient_data.
}
\keyword{datasets}
