% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrival_occupancy.R
\name{occupancy}
\alias{occupancy}
\title{ae_arrival_occupancy_fct}
\usage{
occupancy(date_time, df, start_time = "spell_start",
  end_time = "spell_end")
}
\arguments{
\item{hospital_name}{}
}
\description{
ae_arrival_occupancy_fct
}
