% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupancy_fct.R
\name{occupancy_fct}
\alias{occupancy_fct}
\title{occupancy_fct}
\usage{
occupancy_fct(start_date = as.POSIXct("2016-06-01 00:00:00", tz =
  "Europe/London"), end_date = as.POSIXct("2016-06-08 00:00:00", tz =
  "Europe/London"), data)
}
\arguments{
\item{data}{}
}
\description{
occupancy_fct
}
