% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed_occupancy_on_arrival.R
\name{occupancy_on_arrival}
\alias{occupancy_on_arrival}
\title{occupancy_on_arrival}
\usage{
occupancy_on_arrival(df)
}
\arguments{
\item{df}{dataframe of ED data with start_datetime and end_datetime representing start and end of a patient's time in ED}
}
\value{
the input df with an extra column containing the corresponding occupancy on arrival for each patient
}
\description{
occupancy_on_arrival
}
