% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupancy_weekday_hour.R
\name{occupancy_weekday_hour}
\alias{occupancy_weekday_hour}
\title{occupancy_weekday_hour for A&E}
\usage{
occupancy_weekday_hour(
  start_date = as.POSIXct("2015-04-01 00:00:00"),
  end_date = as.POSIXct("2016-04-01 00:00:00"),
  data,
  plot_chart,
  hospital_name = "Hospital name",
  config_path = "lgt-config/ed"
)
}
\arguments{
\item{config_path}{path to the ed config folder}
}
\value{

}
\description{
occupancy_weekday_hour for A&E
}
