% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_admissions_discharges_day_of_week.R
\name{plot_admissions_discharges_day_of_week}
\alias{plot_admissions_discharges_day_of_week}
\title{Average daily hospital arrivals and discharges}
\usage{
plot_admissions_discharges_day_of_week(
  data,
  startDate,
  endDate,
  returnPlot = TRUE,
  hospitalName = "Hospital name"
)
}
\arguments{
\item{data}{\itemize{Hospital episode data with at least the following fields:
  \item Admissions - date of admission;
  \item Discharges - date of discharge;
  \item IDcol - patient pseudo ID;
  \item PatientType - the admission type, distinguishing emergency from other
  admissions
  \item EpisodeNumber - episode ID;
  \item WardCode - ward where the episode took place.
}}

\item{startDate}{Date of earliest discharge to be included in the analysis.}

\item{endDate}{Date of latest admission to be included in the analysis.}

\item{returnPlot}{If TRUE returns a plot, otherwise returns a dataframe.}

\item{hospitalName}{Name of the hospital.}
}
\value{
A plot (default) or a dataframe showing average admissions and
discharges on each day of the week.
}
\description{
\code{plot_admissions_discharges_day_of_week} returns average daily hospital
arrivals and discharges by days of the week either as a plot (default) or a
dataframe.
}
\examples{
\dontrun{
plot_admissions_discharges_day_of_week(
  data = example_data,
  startDate = "2015-01-01 00:00:00",
  endDate = "2015-02-01 00:00:00",
  returnPlot = TRUE
)
}

}
