% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_average_ed_los_hour_of_day.R
\name{plot_average_ed_los_hour_of_day}
\alias{plot_average_ed_los_hour_of_day}
\title{Average length of stay in the emergency department depending on time of the day}
\usage{
plot_average_ed_los_hour_of_day(
  data,
  startDate,
  endDate,
  returnPlot = TRUE,
  hospitalName = "Hospital name"
)
}
\arguments{
\item{data}{Hospital episode data.}

\item{startDate}{Date of earliest discharge to be included in the analysis.}

\item{endDate}{Date of latest admission to be included in the analysis.}

\item{returnPlot}{If TRUE returns a plot, otherwise returns a dataframe.}

\item{hospitalName}{Name of the hospital.}
}
\value{
A plot (default) or a dataframe showing the average length of stay in
the emergency department of different patient flow groups.
}
\description{
\code{plot_average_ed_los_hour_of_day} allows to visualize the average length
of stay in the emergency department over 24 hour period (0-24h). Patients are
grouped into 4 batches with respect to the flow group they belong to.
}
\examples{
\dontrun{
TBD
}
}
