% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ed_4hourperf_occupancy.R
\name{plot_ed_4hourperf_occupancy}
\alias{plot_ed_4hourperf_occupancy}
\title{Unscheduled emergency department occupancy by weekday and hour}
\usage{
plot_ed_4hourperf_occupancy(
  data,
  startDate,
  endDate,
  returnPlot = TRUE,
  hospitalName = "{hospital_name}"
)
}
\arguments{
\item{data}{A dataframe of with \code{start_datetime} and \code{end_datetime}
representing start and end of a patient's time in the emergency department.}

\item{startDate}{The first date of the period for which the analysis will run.}

\item{endDate}{The last date of the period for which the analysis will run.}

\item{returnPlot}{Plots the graph if set to TRUE, returns a dataframe of the
plot variables if set to FALSE.}

\item{hospitalName}{The name of the hospital for which the analysis is being
done.}
}
\value{
A plot (default) or dataframe showing the proportion of unscheduled
ED attendances with ED LoS >4hrs and wait for treatment >4hrs against ED
occupancy on arrival.
}
\description{
Unscheduled emergency department occupancy by weekday and hour
}
\examples{
\dontrun{
TBD
}

}
