% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ed_4hourperf_timeseries.R
\name{plot_ed_4hourperf_timeseries}
\alias{plot_ed_4hourperf_timeseries}
\title{Emergency department 4 hour performance}
\usage{
plot_ed_4hourperf_timeseries(
  data,
  startDate,
  endDate,
  timeUnit = "day",
  returnPlot = TRUE,
  hospitalName = "{hospital_name}"
)
}
\arguments{
\item{data}{\itemize{Hospital episode data with at least the following fields:
  \item{Attendances - the attendances date and time;}
  \item{Discharges - the discharge date and time;}
  \item{Pseudo id - the patient pseudo id;}
}}

\item{startDate}{A datetime, as a POSIXct, of the earliest attendance to be included in the analysis.}

\item{endDate}{A datetime, as a POSIXct, of the latest attendance to be included in the analysis.}

\item{timeUnit}{The unit of aggregation for 4 hour performance percentages
in the same format as \code{lubridate::round_date}.}

\item{returnPlot}{If TRUE returns a chart, otherwise returns a dataframe.}

\item{hospitalName}{The hospital name for which the analysis is undertaken.}
}
\value{
A plot (default) or a dataframe showing 4 hour performance of the
emergency department.
}
\description{
\code{plot_ed_4hourperf_timeseries} returns daily 4 hour emergency access
performance in the form of a plot or a dataframe.
}
\examples{
\dontrun{
TBD
}
}
