% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ed_4hourperf_timeseries_flow.R
\name{plot_ed_4hourperf_timeseries_flow}
\alias{plot_ed_4hourperf_timeseries_flow}
\title{Emergency department 4 hour performance by flow groups}
\usage{
plot_ed_4hourperf_timeseries_flow(
  data,
  startDate,
  endDate,
  timeUnit = "day",
  returnPlot,
  hospitalName = "{hospital_name}"
)
}
\arguments{
\item{data}{Hospital episode data.}

\item{startDate}{A datetime, as a POSIXct, of the earliest attendance to be included in the analysis.}

\item{endDate}{A datetime, as a POSIXct, of the latest attendance to be included in the analysis.}

\item{timeUnit}{The unit of aggregation for 4 hour performance percentages
in the same format as \code{lubridate::round_date}.}

\item{returnPlot}{If TRUE returns a chart, otherwise returns a dataframe.}

\item{hospitalName}{The hospital name for which the analysis is undertaken.}
}
\value{
A plot (default) or a dataframe showing 4 hour performance of the
emergency department in each of the 4 flow groups.
}
\description{
\code{plot_ed_4hourperf_timeseries_flow} returns daily 4 hour emergency access
performance in each of the 4 flow groups in the form of a plot or a dataframe.
}
\examples{
\dontrun{
TBD
}
}
