% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ed_arrival_occupancy_hour_of_day.R
\name{plot_ed_arrival_occupancy_hour_of_day}
\alias{plot_ed_arrival_occupancy_hour_of_day}
\title{Average emergency department arrival occupancy during the day}
\usage{
plot_ed_arrival_occupancy_hour_of_day(
  data,
  startDate,
  endDate,
  returnPlot = TRUE,
  hospitalName = "Hospital name"
)
}
\arguments{
\item{data}{\itemize{Hospital episode data with at least the following fields:
  \item Admissions - date of admission;
  \item Discharges - date of discharge;
  \item IDcol - patient pseudo ID;
  \item PatientType - the admission type, distinguishing emergency from other
  admissions
  \item EpisodeNumber - episode ID;
  \item WardCode - ward where the episode took place.
}}

\item{startDate}{Date of earliest discharge to be included in the analysis.}

\item{endDate}{Date of latest admission to be included in the analysis.}

\item{returnPlot}{If TRUE returns a plot, otherwise returns a dataframe.}

\item{hospitalName}{Name of the hospital.}
}
\value{
A plot (default) or a dataframe showing average occupancy of emergency
department over a day.
}
\description{
\code{plot_ed_arrival_occupancy_hour_of_day} returns average daily occupancy
of the emergency department by hours of the day either as a plot (default) or
a dataframe.
}
\examples{
TBD
}
