% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ed_attendance_demographics.R
\name{plot_ed_attendance_demographics}
\alias{plot_ed_attendance_demographics}
\title{Attendance and admissions by age and gender in emergency department (ED)}
\usage{
plot_ed_attendance_demographics(
  data,
  startDate,
  endDate,
  returnPlot = TRUE,
  hospitalName = "Hospital name"
)
}
\arguments{
\item{data}{Hospital episode data.}

\item{startDate}{Date of earliest discharge to be included in the analysis.}

\item{endDate}{Date of latest admission to be included in the analysis.}

\item{returnPlot}{If TRUE returns a plot, otherwise returns a dataframe.}

\item{hospitalName}{Name of the hospital.}
}
\value{
A plot (default) or a dataframe showing attendance of emergency
department and admissions therefrom by age and gender.
}
\description{
\code{plot_ed_attendance_demographics} returns the number of ED attendances
and hospitalizations by age (divided into bands) and gender (male and female).
}
\examples{
\dontrun{
plot_ed_attendance_demographics(
  data = test_data_age_sex_att_adm,
  startDate = "2012-01-01 00:00:00",
  endDate = "2015-01-01 00:00:00",
  returnPlot = TRUE
)
}
}
