% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ed_attendances_timeseries_flow.R
\name{plot_ed_attendances_timeseries_flow}
\alias{plot_ed_attendances_timeseries_flow}
\title{Weekly attendance of emergency department by flow groups}
\usage{
plot_ed_attendances_timeseries_flow(
  data,
  startDate,
  endDate,
  timeUnit = "week",
  returnPlot = TRUE,
  hospitalName = "{hospital_name}",
  restrictPlotRange = TRUE
)
}
\arguments{
\item{data}{Hospital episode data.}

\item{startDate}{The first date of the period for which the analysis will run.}

\item{endDate}{The last date of the period for which the analysis will run.}

\item{timeUnit}{TBD}

\item{returnPlot}{Plots the graph if set to TRUE, returns a dataframe otherwise.}

\item{hospitalName}{The name of the hospital for which the analysis is being
done.}

\item{restrictPlotRange}{A boolean indicating whether to zoom the plot to only
the range specified by \code{startDate} and \code{endDate}.}
}
\value{
A lineplot (default) or a dataframe showing the distribution of ED
attendances for different flow groups.
}
\description{
\code{plot_ed_attendances_timeseries_flow} shows weekly attendance of emergency
department by different flow groups.
}
\examples{
\dontrun{
TBD
}
}
