% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ed_los_distribution_admission.R
\name{plot_ed_los_distribution_admission}
\alias{plot_ed_los_distribution_admission}
\title{Length of stay inthe emergency department for attendances and admissions}
\usage{
plot_ed_los_distribution_admission(
  data,
  startDate,
  endDate,
  returnPlot,
  hospitalName
)
}
\arguments{
\item{data}{Hospital episode data.}

\item{startDate}{Date of earliest discharge to be included in the analysis.}

\item{endDate}{Date of latest admission to be included in the analysis.}

\item{returnPlot}{If TRUE returns a plot, otherwise returns a dataframe.}

\item{hospitalName}{Name of the hospital.}
}
\value{
A plot (default) or a dataframe showing the distribution of length of
stay for patients in the emergency department.
}
\description{
\code{plot_ed_los_distribution_admission} shows the distribution of length of
stay for people in the emergency department, stratified by whether they were
admitted to the hospital or discharged.
}
\examples{
\dontrun{
TBD
}
}
