% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ed_los_distribution_flow.R
\name{plot_ed_los_distribution_flow}
\alias{plot_ed_los_distribution_flow}
\title{Distribution of time spent in the emergency department by flow grow}
\usage{
plot_ed_los_distribution_flow(
  data,
  startDate,
  endDate,
  returnPlot,
  hospitalName = "{hospital_name}"
)
}
\arguments{
\item{data}{Hospital episode data.}

\item{startDate}{Date of earliest discharge to be included in the analysis.}

\item{endDate}{Date of latest admission to be included in the analysis.}

\item{returnPlot}{If TRUE returns a plot, otherwise returns a dataframe.}

\item{hospitalName}{Name of the hospital.}
}
\value{
A plot (default) or a dataframe containign the information on the
distribution of the length of stay in the emergency department.
}
\description{
\code{plot_ed_los_distribution_flow} gives a distribution of the length of
stay in the emergency department for each of the 4 flow groups.
}
\examples{
\dontrun{
TBD
}
}
