% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ed_occupancy_hour_of_week.R
\name{plot_ed_occupancy_hour_of_week}
\alias{plot_ed_occupancy_hour_of_week}
\title{Unscheduled emergency department occupancy by weekday and hour}
\usage{
plot_ed_occupancy_hour_of_week(
  data,
  startDate,
  endDate,
  returnPlot = TRUE,
  hospitalName = "{hospital_name}"
)
}
\arguments{
\item{data}{Hospital episode data.}

\item{startDate}{Date of earliest discharge to be included in the analysis.}

\item{endDate}{Date of latest admission to be included in the analysis.}

\item{returnPlot}{If TRUE returns a plot, otherwise returns a dataframe.}

\item{hospitalName}{Name of the hospital.}
}
\value{
A plot (default) or a dataframe showing average emergency department
occupancy by weekday and hour.
}
\description{
\code{plot_ed_occupancy_hour_of_week} shows average occupancy in an emergency
department throughout weekdays (Mon-Sun).
}
\examples{
\dontrun{
TBD
}
}
