% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey_flow.R
\name{plot_flow_diagram}
\alias{plot_flow_diagram}
\title{plot_flow_diagram}
\usage{
plot_flow_diagram(
  moves_table,
  ward_level = F,
  start = NULL,
  end = NULL,
  selected_levels = NULL,
  remove_static_moves = F
)
}
\arguments{
\item{moves_table}{table of patient episode moves within and outside the hospital}

\item{ward_level}{boolean of whether the analysis is done ward by ward or by
hospital category (e.g. Mecical, Surgical, Acute etc.)}

\item{start}{the datetime from which the moves should be counted}

\item{end}{the datetime from which the moves should stop being counted}

\item{selected_levels}{a vector of strings stating the wards or categories
that should be included in the analysis
e.g. c("External Incoming", "Medical", "Surgical", "External Outgoing")}
}
\value{
spell table
}
\description{
plot_flow_diagram
}
