% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_los_distribution_admission_method.R
\name{plot_los_distribution_admission_method}
\alias{plot_los_distribution_admission_method}
\title{Length of stay distribution for hospitalized patients}
\usage{
plot_los_distribution_admission_method(
  data,
  startDate,
  endDate,
  returnPlot,
  hospitalName = "Hospital Name"
)
}
\arguments{
\item{data}{Hospital episode data.}

\item{startDate}{Date of earliest discharge to be included in the analysis.}

\item{endDate}{Date of latest admission to be included in the analysis.}

\item{returnPlot}{If TRUE returns a plot, otherwise returns a dataframe.}

\item{hospitalName}{Name of the hospital.}
}
\value{
A plot (default) or a dataframe showing the distribution of length of
stay for hospitalized patients.
}
\description{
\code{plot_los_distribution_admission_method} shows the distribution of length of
stay for people in the inpatient department, stratified by the type of discharge.
}
\examples{
\dontrun{
TBD
}
}
