% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_readmissions_timeseries.R
\name{plot_readmissions_timeseries}
\alias{plot_readmissions_timeseries}
\title{Readmission frequency over period of time}
\usage{
plot_readmissions_timeseries(
  data,
  startDate,
  endDate,
  readmissionBy,
  returnPlot = TRUE,
  hospitalName = "{hospital_name}"
)
}
\arguments{
\item{data}{Hospital episode data.}

\item{startDate}{The first date of the period for which the analysis will run.}

\item{endDate}{The last date of the period for which the analysis will run.}

\item{readmissionBy}{Integer indicating readmission timeframe.}

\item{returnPlot}{Plots the graph if set to TRUE, returns a dataframe otherwise.}

\item{hospitalName}{The name of the hospital for which the analysis is being
done.}
}
\value{
A lineplot (default) or a dataframe showing the proportion of readmissions
over provided period of time.
}
\description{
\code{plot_readmissions_timeseries} shows readmission percentage over period
of time. Indicates violation of Schewart rules as well.
}
\examples{
\dontrun{
TBD
}
}
