% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_modelling.R
\name{predict_residual_occupancy}
\alias{predict_residual_occupancy}
\title{predict_residual_occupancy}
\usage{
predict_residual_occupancy(df, t, S_function, delta_t, coxmodel = NULL)
}
\arguments{
\item{delta_t}{}
}
\value{

}
\description{
predict_residual_occupancy
}
