% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_standardise.R
\name{recode_factors}
\alias{recode_factors}
\title{recode_factors}
\usage{
recode_factors(provided_data, config_path)
}
\arguments{
\item{provided_data}{a tibble containing the imported dataset included all standard columns parsed
as the correct type}

\item{config_path}{path to the directory containing the config files for this dataset}
}
\value{
copy of provided_data with factor variables recoded to standard levels as
per the config files
}
\description{
recode_factors
}
