% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_example_report.R
\name{render_example_report}
\alias{render_example_report}
\title{render_example_report}
\usage{
render_example_report(
  sample_weeks = 4,
  timezone_config_path = "../lgt-config/ed"
)
}
\arguments{
\item{sample_weeks}{The number of weeks of sample data to use for the report.}
}
\value{
The system.time output for the time taken to render the report.
}
\description{
render_example_report
}
