% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_modelling.R
\name{run_S_func_model}
\alias{run_S_func_model}
\title{run_S_func_model}
\usage{
run_S_func_model(df, S_func, date_seq, horizon = 48)
}
\arguments{
\item{df}{dataset of spells}

\item{S_func}{survival function}

\item{date_seq}{sequence of dates to predict for}

\item{horizon}{how far ahead to predict (default 48h)}
}
\value{
object containing prediction table and diagnostics
}
\description{
run_S_func_model
}
