% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_dataset.R
\name{spell_intersections}
\alias{spell_intersections}
\title{spell_intersections}
\usage{
spell_intersections(data1, data2)
}
\arguments{
\item{data1}{a dataset with a column called 'Spell Number'}

\item{data2}{a dataset with a column called 'Spell Number'}
}
\value{
list of three vectors giving distinct values of spell number present in both, and only in each,
of data1 and data2
}
\description{
spell_intersections
}
