% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_dataset.R
\name{spell_type}
\alias{spell_type}
\title{spell_type}
\usage{
spell_type(spell_episode_types)
}
\arguments{
\item{spell_episode_types}{a vector containing the episode types for episodes
within one spell in hospital.}
}
\value{
the type of the spell - if all episode types are equal, this unique type is the
type of the spell, otherwise the type of the spell is 'Mixed'.
}
\description{
spell_type
}
