% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell_linkage.R
\name{spell_variables}
\alias{spell_variables}
\title{spell_variables}
\usage{
spell_variables(all_episodes)
}
\arguments{
\item{all_episodes}{all ED and Inpatient episodes, with spell_number.
This should be the return value of a call to make_spell_number}
}
\value{
the spell table
}
\description{
spell_variables
}
