% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_dataset.R
\name{standardise_column_names}
\alias{standardise_column_names}
\title{standardise_colum_names}
\usage{
standardise_column_names(data, colname_mapping)
}
\arguments{
\item{data}{episode data whose column names are to be standardised}

\item{colname_mapping}{a dataframe with two columns, dataCol - the old names,
and analysisCol - their new equivalents}
}
\value{
data, with column names standardised according to colname_mapping
}
\description{
standardise_colum_names
}
\examples{
\dontrun{
standardise_column_names(data = dataset, colname_mapping = lgt_col_mapping())
}
}
