% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed_occupancy_on_arrival.R
\name{wait_times_over_4hrs_Vs_occupancy}
\alias{wait_times_over_4hrs_Vs_occupancy}
\title{wait_times_over_4hrs_Vs_occupancy}
\usage{
wait_times_over_4hrs_Vs_occupancy(df, startDate, endDate, hospitalName,
  plotChart = T)
}
\arguments{
\item{df}{dataframe of ED data with start_datetime and end_datetime representing start and end of a patient's time in ED}

\item{startDate}{the first date of the period for which the analysis will run}

\item{endDate}{the last date of the period for which the analysis will run}

\item{hospitalName}{the name of the hospital for which the analysis is being done}

\item{plotChart}{plots the graph if set to TRUE, returns dataframe of the plot variables if set to FALSE}
}
\value{
graph or dataframe of % of unscheduled ED attendances with ED LoS >4hrs and wait for treatment >4hrs
against ED occupancy on arrival
}
\description{
wait_times_over_4hrs_Vs_occupancy
}
