% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly_ed_att_flgrp.R
\name{weekly_ed_att_flgrp}
\alias{weekly_ed_att_flgrp}
\title{weekly_ed_att_flgrp}
\usage{
weekly_ed_att_flgrp(start_date, end_date, data, time_unit = "weekly",
  plot_chart, hospital_name)
}
\arguments{
\item{hospital_name}{}
}
\description{
weekly_ed_att_flgrp
}
