% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hf_utilities.R
\name{add_directorate_variable}
\alias{add_directorate_variable}
\title{add_directorate_variable}
\usage{
add_directorate_variable(spell_table)
}
\arguments{
\item{spell_table}{}
}
\description{
add_directorate_variable
}
