% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell_linkage.R
\name{add_spell_variables}
\alias{add_spell_variables}
\title{add_spell_variables}
\usage{
add_spell_variables(ed_data, inpatient_data, spell_table)
}
\arguments{
\item{spell_table}{pre-created from ed_data and inpatient_data using make_spell_table}
}
\value{

}
\description{
add_spell_variables
}
