% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrival_occupancy.R
\name{ae_arrival_occupancy}
\alias{ae_arrival_occupancy}
\title{ae_arrival_occupancy}
\usage{
ae_arrival_occupancy(
  start_date,
  end_date,
  data,
  plot_chart,
  hospital_name = "Hospital name"
)
}
\arguments{
\item{hospital_name}{}
}
\value{

}
\description{
ae_arrival_occupancy
}
