% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrival_occupancy.R
\name{ae_arrival_occupancy}
\alias{ae_arrival_occupancy}
\title{ae_arrival_occupancy}
\usage{
ae_arrival_occupancy(
  start_date = as.POSIXct("2012-01-01 00:00:00"),
  end_date = as.POSIXct("2015-01-01 00:00:00"),
  data,
  plot_chart,
  hospital_name = "Hospital name",
  config_path = "lgt-config/ed"
)
}
\arguments{
\item{config_path}{path to the ed config folder}
}
\value{

}
\description{
ae_arrival_occupancy
}
