% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_ed_los_flows.R
\name{average_ed_los_flows}
\alias{average_ed_los_flows}
\title{Title}
\usage{
average_ed_los_flows(
  start_date,
  end_date,
  data,
  plot_chart,
  hospital_name = "Hospital name"
)
}
\arguments{
\item{hospital_name}{}
}
\value{

}
\description{
Title
}
