% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four_hrs_perf_flow_groups.R
\name{format_control_chart}
\alias{format_control_chart}
\title{format_control_chart}
\usage{
format_control_chart(cht, r1_col, r2_col, main_col = "black")
}
\arguments{
\item{cht}{ggplot chart object}

\item{r1_col}{colour to highlight rule 1 breaks}

\item{r2_col}{colour to highlight rule 2 breaks}

\item{main_col}{colour without highlight}
}
\value{
chart with formatting applied
}
\description{
format_control_chart
}
