% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four_hrs_perf_flow_groups.R
\name{four_hrs_perf_flow_groups}
\alias{four_hrs_perf_flow_groups}
\title{four hrs perf flow_groups}
\usage{
four_hrs_perf_flow_groups(start_dt = as.Date("2016-01-01", tz =
  "Europe/London"), end_dt = as.Date("2016-09-01", tz = "Europe/London"),
  data, time_unit = "day", plot_chart, hospital_name = "Hospital_Name")
}
\arguments{
\item{hospital_name}{}
}
\description{
four hrs perf flow_groups
}
