% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_modelling.R
\name{get_discharge_probability}
\alias{get_discharge_probability}
\title{get_discharge_probability}
\usage{
get_discharge_probability(time_in, KM, delta_t)
}
\arguments{
\item{time_in}{the length of time elapsed so far}

\item{delta_t}{time interval}
}
\value{
probability still in after additional delta_t
}
\description{
get_discharge_probability
}
