% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_standardise.R
\name{import_and_standardise}
\alias{import_and_standardise}
\title{import_and_standardise}
\usage{
import_and_standardise(
  data_import_list,
  remove_duplicates = TRUE,
  tz_config_path = "../lgt-config/ed"
)
}
\arguments{
\item{data_import_list}{list of named lists, each named list has four elements called
data_path, config_path, site, and facility. The first two of these are character strings specifying the paths
to i) a csv file to import data from and ii) a folder containing hospitalflow config
files. The latter two specify the site and facility (e.g. ED, IP, UCC).}

\item{remove_duplicates}{boolean to control whether data is de-duped (TRUE) or not (FALSE).}

\item{tz_config_path}{path to the ed config folder}
}
\value{
list of tibbles, each containing standardised import of one data file from data_paths
}
\description{
import_and_standardise
}
