% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/los_distrib_method_admission.R
\name{los_distrib_method_admission}
\alias{los_distrib_method_admission}
\title{los_distrib_method_admission}
\usage{
los_distrib_method_admission(
  start_date = as.Date("2012-01-01"),
  end_date = as.Date("2015-01-01"),
  data,
  plot_chart,
  hospital_name = "Hospital Name",
  config_path = "lgt-config/ed"
)
}
\arguments{
\item{config_path}{path to the ed config folder}
}
\value{

}
\description{
los_distrib_method_admission
}
