% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves_table.R
\name{make_moves_table}
\alias{make_moves_table}
\title{make_moves_table}
\usage{
make_moves_table(
  ed_data = test_ed_data_sample,
  inpatient_data = test_ip_data_sample
)
}
\arguments{
\item{ed_data}{standard ED data}

\item{inpatient_data}{standard inpatient data}
}
\value{
moves table
}
\description{
make_moves_table
}
