% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell_linkage.R
\name{make_spell_table}
\alias{make_spell_table}
\title{make_spell_table}
\usage{
make_spell_table(
  ed_data,
  inpatient_data,
  same_type_episode_lag = 1,
  different_type_episode_lag = 6
)
}
\arguments{
\item{ed_data}{standard ED data}

\item{inpatient_data}{standard inpatient data}

\item{same_type_episode_lag}{the maximum amount of time two episodes of the same type
can be separated by and still be classified as part of the same spell.}

\item{different_type_episode_lag}{the maximum amount of time two episodes of different type
can be separated by and still be classified as part of the same spell.}
}
\value{
spell table
}
\description{
make_spell_table
}
