% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_timeser.R
\name{mortality_timeser}
\alias{mortality_timeser}
\title{mortality_timeser}
\usage{
mortality_timeser(
  start_date = as.POSIXct("2016-01-01 00:00:00", tz = "Europe/London"),
  end_date = as.POSIXct("2016-03-31 00:00:00", tz = "Europe/London"),
  data,
  plot_chart,
  hospital_name
)
}
\arguments{
\item{hospital_name}{}
}
\value{

}
\description{
mortality_timeser
}
