% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_timeseries.R
\name{mortality_timeseries}
\alias{mortality_timeseries}
\title{mortality_timeseries}
\usage{
mortality_timeseries(start_date, end_date, data, plot_chart, hospital_name)
}
\arguments{
\item{hospital_name}{}
}
\value{

}
\description{
mortality_timeseries
}
