% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupancy_fct.R
\name{occupancy}
\alias{occupancy}
\title{occupancy}
\usage{
occupancy(time_instance, df, df_type = "spell table")
}
\arguments{
\item{time_instance}{datetime at which we want to know the occupancy}

\item{df}{dataframe on which to perform the occupancy function}

\item{df_type}{whether the df specified is a "spell table" or an "inpatient table" or "ED table"}
}
\value{
the occupancy at date_time
}
\description{
occupancy
}
