% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupancy_fct.R
\name{occupancy}
\alias{occupancy}
\title{occupancy}
\usage{
occupancy(date_time, df, start_time = "spell_start",
  end_time = "spell_end")
}
\arguments{
\item{date_time}{datetime at which we want to know the occupancy}

\item{df}{dataframe representing stays, with start_datetime and end_datetime specifying stay start and end}
}
\value{
the occupancy at date_time
}
\description{
occupancy
}
