% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ed_4hourperf_occupancy.R
\name{occupancy_on_arrival}
\alias{occupancy_on_arrival}
\title{Emergency department occupancy on arrival}
\usage{
occupancy_on_arrival(data)
}
\arguments{
\item{data}{A dataframe of with \code{start_datetime} and \code{end_datetime}
representing start and end of a patient's time in the emergency department.}
}
\value{
The input \code{data} with an extra column \code{occupancy_on_arrival}.
}
\description{
Computes the emergency department occupancy on arrival for each patient.
}
\examples{
\dontrun{
TBD
}

}
