% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmissions_ip.R
\name{readmissions_ip}
\alias{readmissions_ip}
\title{readmissions_ip}
\usage{
readmissions_ip(
  start_date,
  end_date,
  data,
  readmission_by,
  plot_chart,
  hospital_name
)
}
\arguments{
\item{readmission_by}{}
}
\value{

}
\description{
readmissions_ip
}
