% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmissions_ip.R
\name{readmissions_ip}
\alias{readmissions_ip}
\title{readmissions_ip}
\usage{
readmissions_ip(start_date = as.POSIXct("2016-01-01 00:00:00", tz =
  "Europe/London"), end_date = as.POSIXct("2016-03-31 00:00:00", tz =
  "Europe/London"), data, readmission_by, plot_chart, hospital_name)
}
\arguments{
\item{readmission_by}{}
}
\description{
readmissions_ip
}
